+++ src/app/helpers/index.ts
@@ -2,3 +2,4 @@ export * from './command.helper';
 export * from './dialog.helper';
 export * from './filesystem.helper';
 export * from './inflector.helper';
+export * from './placeholder.helper';